Upgrade to tdlib 1.7.9 and fixup

See merge request ubports/apps/teleports!420

+++ libs/qtdlib/chat/requests/qtdgetsecretchatrequest.cpp
@@ -7,7 +7,7 @@ QTdGetSecretChatRequest::QTdGetSecretChatRequest(QObject *parent)
 {
 }
 
-void QTdGetSecretChatRequest::setSecretChatId(const qint32 &id)
+void QTdGetSecretChatRequest::setSecretChatId(const qint64 &id)
 {
     m_chatId = id;
 }
